[#4] - Fixing details of to ensure the customer will recieve the right data
