+++ src/main/scala/com/thebrauproject/creation/EventRouter.scala
@@ -2,11 +2,10 @@ package com.thebrauproject.creation
 
 import akka.actor.{Actor, ActorLogging, Props}
 import akka.pattern.ask
-import com.thebrauproject.operations.OperationsDb._
-import com.thebrauproject.operations.OperationsDb.Connect
+import com.thebrauproject.operations._
 
-class DatabaseRouter extends Actor with ActorLogging {
-  val databaseOperation = context.actorOf(Props[DatabaseActor], "DatabaseHero")
+class EventRouter extends Actor with ActorLogging {
+  val databaseOperation = context.actorOf(Props[EventActor], "DatabaseHero")
   override def receive: Receive = {
     case op: DBCreatureOperation =>
       log.info(s"Operation with Creature")
