+++ src/main/scala/com/thebrauproject/kafka/HeroConsumer.scala
@@ -2,28 +2,35 @@ package com.thebrauproject.kafka
 
 import java.util.Properties
 
-import akka.actor.{Actor, ActorSystem, Props}
+import akka.actor.{Actor, ActorRef, ActorSystem, Props}
+import com.thebrauproject.creation.CreateHero
+import com.thebrauproject.elements.Hero
 import com.thebrauproject.modification.RedisDbActor
-import com.thebrauproject.operations.OperationsDb.{Connect, RedisObject}
-import com.thebrauproject.operations.OperationsKafka.StartConsumer
+import com.thebrauproject.operations._
+import com.thebrauproject.elements.implicits._
 import org.apache.kafka.common.serialization.Serdes
 import org.apache.kafka.streams.{KafkaStreams, StreamsBuilder, StreamsConfig}
 import org.apache.kafka.streams.kstream.KStream
+import spray.json.DefaultJsonProtocol._
+import spray.json._
 
 
 class HeroConsumer extends Actor {
   val system = ActorSystem("HeroConsumer")
 
-  val redisDB = system.actorOf(Props[RedisDbActor], "dbCreator")
+  val redisDB: ActorRef = system.actorOf(Props[RedisDbActor], "dbRedisCreator")
+  val postgresDB: ActorRef = system.actorOf(Props[CreateHero], "dbPostgresCreator")
 
   val builder = new StreamsBuilder()
 
   val heroes: KStream[String, String] = builder.stream("hero")
 
   redisDB ! Connect
+  postgresDB ! Connect
 
   heroes.foreach((k, v) => {
     redisDB ! RedisObject(k, v)
+    postgresDB ! v.parseJson.convertTo[Hero]
   })
 
   override def receive: Receive = {
