--- src/main/scala/com/thebrauproject/operations/OperationsKafka.scala
@@ -1,8 +1,0 @@
-package com.thebrauproject.operations
-
-object OperationsKafka {
-
-  case object StartConsumer
-
-  case object StartProducer
-}
